solved Euler problem 41

+++ java_challenges/main/Prime.java
@@ -2,8 +2,8 @@ import java.util.Collections;
 import java.util.HashSet;
 
 public class Prime {
-    private static Integer last = 2;
-    private static HashSet<Integer> primes = new HashSet<>(Collections.singletonList(2));
+    private static long last = 2L;
+    private static HashSet<Long> primes = new HashSet<>(Collections.singletonList(2L));
 
     public static Boolean isPrime(long tested) {
         if (primes.contains(tested)) return true;
@@ -16,7 +16,7 @@ public class Prime {
     }
 
     private static boolean lastIsPrime() {
-        for (Integer i = 2; i <= Math.sqrt(last); i++) {
+        for (long i = 2; i <= Math.sqrt(last); i++) {
             if (primes.contains(i) && last % i == 0) return false;
         }
         return true;
